Reformat: Typecast Unsigned Long int to an int

+++ 0x06-pointers_arrays_strings/6-cap_string.c
@@ -20,7 +20,7 @@ char *cap_string(char *s)
 
 	for (i = 0; s[i] != '\0'; i++)
 	{
-		for (j = 0; j < strlen(caps); j++)
+		for (j = 0; j < (int)strlen(caps); j++)
 		{
 			if (s[i] == lower[j] && i == 0)
 			{
